+++ src/input.rs
@@ -12,7 +12,7 @@ use std::{
 };
 
 static FAST_POLL_DURATION: Duration = Duration::from_millis(100);
-static SLOW_POLL_DURATION: Duration = Duration::from_millis(1000);
+static SLOW_POLL_DURATION: Duration = Duration::from_millis(10000);
 
 ///
 #[derive(Clone, Copy, Debug)]
@@ -123,6 +123,9 @@ impl Input {
 					}
 
 					tx.send(InputEvent::Input(e))?;
+					//Note: right after an input event we might have a reason to stop
+					// polling (external editor opening) so lets do a quick poll until the next input
+					// this fixes https://github.com/extrawurst/gitui/issues/1506
 					poll_duration = FAST_POLL_DURATION;
 				} else {
 					poll_duration = SLOW_POLL_DURATION;